environement/contribution_ae/requirements.txt

asgiref==3.7.2
Django==4.2.1
django-cors-headers==4.2.0
django-dev-email==1.0.2
django-discover-runner==1.0
django-role-permissions==3.1.1
#dotenv==0.0.5
fpdf==1.7.2
#psycopg2==2.9.6
python-dotenv==0.21.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
Unidecode==1.3.4
XlsxWriter==3.1.2
